+++ crates/forge/tests/e2e/build_trace_data.rs
@@ -16,24 +16,24 @@ fn simple_package_save_trace() {
 
     assert!(temp
         .join(TRACE_DIR)
-        .join("simple_package::tests::test_fib.json")
+        .join("simple_package_tests_test_fib.json")
         .exists());
     assert!(!temp
         .join(TRACE_DIR)
-        .join("simple_package_integrationtest::test_simple::test_failing.json")
+        .join("simple_package_integrationtest_test_simple_test_failing.json")
         .exists());
     assert!(!temp
         .join(TRACE_DIR)
-        .join("simple_package::tests::ignored_test.json")
+        .join("simple_package_tests_ignored_test.json")
         .exists());
     assert!(temp
         .join(TRACE_DIR)
-        .join("simple_package_integrationtest::ext_function_test::test_simple.json")
+        .join("simple_package_integrationtest_ext_function_test_test_simple.json")
         .exists());
 
     let trace_data = fs::read_to_string(
         temp.join(TRACE_DIR)
-            .join("simple_package_integrationtest::ext_function_test::test_simple.json"),
+            .join("simple_package_integrationtest_ext_function_test_test_simple.json"),
     )
     .unwrap();
 
@@ -56,7 +56,7 @@ fn trace_has_contract_and_function_names() {
 
     let trace_data = fs::read_to_string(
         temp.join(TRACE_DIR)
-            .join("trace_info_integrationtest::test_trace::test_trace.json"),
+            .join("trace_info_integrationtest_test_trace_test_trace.json"),
     )
     .unwrap();
 
@@ -103,7 +103,7 @@ fn trace_has_cairo_execution_info() {
 
     let trace_data = fs::read_to_string(
         temp.join(TRACE_DIR)
-            .join("trace_info_integrationtest::test_trace::test_trace.json"),
+            .join("trace_info_integrationtest_test_trace_test_trace.json"),
     )
     .unwrap();
 
@@ -145,7 +145,7 @@ fn trace_has_deploy_with_no_constructor_phantom_nodes() {
 
     let trace_data = fs::read_to_string(
         temp.join(TRACE_DIR)
-            .join("trace_info_integrationtest::test_trace::test_trace.json"),
+            .join("trace_info_integrationtest_test_trace_test_trace.json"),
     )
     .unwrap();
 
@@ -165,4 +165,33 @@ fn trace_has_deploy_with_no_constructor_phantom_nodes() {
         call_trace.nested_calls[2],
         cairo_annotations::trace_data::CallTraceNode::DeployWithoutConstructor
     );
+}
+
+#[test]
+fn trace_is_produced_even_if_contract_panics() {
+    let temp = setup_package("backtrace_panic");
+    test_runner(&temp)
+        .arg("--save-trace-data")
+        .assert()
+        .success();
+
+    let trace_data = fs::read_to_string(
+        temp.join(TRACE_DIR)
+            .join("backtrace_panic_Test_test_contract_panics.json"),
+    )
+    .unwrap();
+
+    let call_trace: ProfilerCallTrace = serde_json::from_str(&trace_data).unwrap();
+
+    assert_all_execution_info_exists(&call_trace);
+}
+
+fn assert_all_execution_info_exists(trace: &ProfilerCallTrace) {
+    assert!(trace.cairo_execution_info.is_some());
+
+    for trace_node in &trace.nested_calls {
+        if let ProfilerCallTraceNode::EntryPointCall(trace) = trace_node {
+            assert_all_execution_info_exists(trace);
+        }
+    }
 }