+++ crates/forge/tests/e2e/mod.rs
@@ -11,10 +11,13 @@ mod coverage;
 mod docs_snippets_validation;
 mod env;
 mod features;
+#[cfg(not(target_os = "windows"))]
 mod fork_warning;
+#[cfg(not(target_os = "windows"))]
 mod forking;
 mod fuzzing;
 mod io_operations;
+mod requirements;
 mod running;
 mod steps;
 mod trace_print;