+++ crates/sncast/src/helpers/mod.rs
@@ -5,5 +5,7 @@ pub mod configuration;
 pub mod constants;
 pub mod error;
 pub mod fee;
+pub mod interactive;
 pub mod rpc;
 pub mod scarb_utils;
+pub mod version;