+++ crates/sncast/tests/e2e/account/import.rs
@@ -275,6 +275,43 @@ pub async fn test_happy_case_add_profile() {
 }
 
 #[tokio::test]
+pub async fn test_add_profile_with_network_arg() {
+    let tempdir = tempdir().expect("Failed to create a temporary directory");
+    let accounts_file = "accounts.json";
+
+    let args = vec![
+        "--accounts-file",
+        accounts_file,
+        "account",
+        "import",
+        "--network",
+        "sepolia",
+        "--name",
+        "my_account_import",
+        "--address",
+        "0x1",
+        "--private-key",
+        "0x2",
+        "--class-hash",
+        DEVNET_OZ_CLASS_HASH_CAIRO_0,
+        "--type",
+        "oz",
+        "--add-profile",
+        "my_account_import",
+    ];
+
+    let snapbox = runner(&args).current_dir(tempdir.path());
+    let output = snapbox.assert();
+
+    assert_stderr_contains(
+        output,
+        indoc! {r"
+        error: the argument '--network <NETWORK>' cannot be used with '--add-profile <ADD_PROFILE>'
+    "},
+    );
+}
+
+#[tokio::test]
 pub async fn test_detect_deployed() {
     let tempdir = tempdir().expect("Unable to create a temporary directory");
     let accounts_file = "accounts.json";
@@ -451,12 +488,18 @@ pub async fn test_invalid_private_key_file_path() {
     let snapbox = runner(&args);
     let output = snapbox.assert().success();
 
+    let expected_file_error = if cfg!(target_os = "windows") {
+        "The system cannot find the file specified[..]"
+    } else {
+        "No such file or directory [..]"
+    };
+
     assert_stderr_contains(
         output,
-        indoc! {r"
+        formatdoc! {r"
         command: account import
-        error: Failed to obtain private key from the file my_private_key: No such file or directory (os error 2)
-        "},
+        error: Failed to obtain private key from the file my_private_key: {}
+        ", expected_file_error},
     );
 }
 
@@ -713,7 +756,7 @@ pub async fn test_happy_case_default_name_generation() {
         "delete",
         "--name",
         "account-2",
-        "--network",
+        "--network-name",
         "alpha-sepolia",
     ];
 
@@ -740,7 +783,7 @@ pub async fn test_happy_case_default_name_generation() {
         let snapbox = runner(&import_args).current_dir(tempdir.path());
         snapbox.assert().stdout_matches(formatdoc! {r"
         command: account import
-        account_name: Account imported with name: account-{id}
+        account_name: account-{id}
         add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
     ", id = i + 1});
     }
@@ -766,7 +809,7 @@ pub async fn test_happy_case_default_name_generation() {
     let snapbox = runner(&import_args).current_dir(tempdir.path());
     snapbox.assert().stdout_matches(indoc! {r"
         command: account import
-        account_name: Account imported with name: account-2
+        account_name: account-2
         add_profile: --add-profile flag was not set. No profile added to snfoundry.toml
     "});